+++ packages/sdks/react-sdk/src/components/ApplicationsPortal.tsx
@@ -24,7 +24,7 @@ const ApplicationsPortal = React.forwardRef<
 
   useImperativeHandle(ref, () => innerRef);
 
-  const { projectId, baseUrl, baseStaticUrl, baseCdnUrl } =
+  const { projectId, baseUrl, baseStaticUrl, baseCdnUrl, refreshCookieName } =
     React.useContext(Context);
 
   return (
@@ -41,6 +41,7 @@ const ApplicationsPortal = React.forwardRef<
           'theme.attr': theme,
           'debug.attr': debug,
           'styleId.attr': styleId,
+          'refreshCookieName.attr': refreshCookieName,
           // props
           'logger.prop': logger,
         }}