+++ packages/sdks/angular-sdk/CHANGELOG.md
@@ -2,6 +2,49 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.9.5](https://github.com/descope/descope-js/compare/angular-sdk-0.9.4...angular-sdk-0.9.5) (2025-02-26)
+
+### Dependency Updates
+
+* `access-key-management-widget` updated to version `0.4.1`
+* `audit-management-widget` updated to version `0.4.1`
+* `role-management-widget` updated to version `0.3.1`
+* `user-management-widget` updated to version `0.7.1`
+* `user-profile-widget` updated to version `0.3.2`
+* `applications-portal-widget` updated to version `0.3.1`
+* `web-component` updated to version `3.37.0`
+## [0.9.4](https://github.com/descope/descope-js/compare/angular-sdk-0.9.3...angular-sdk-0.9.4) (2025-02-25)
+
+### Dependency Updates
+
+* `user-profile-widget` updated to version `0.3.1`
+* `web-component` updated to version `3.36.1`
+## [0.9.3](https://github.com/descope/descope-js/compare/angular-sdk-0.9.2...angular-sdk-0.9.3) (2025-02-24)
+
+### Dependency Updates
+
+* `access-key-management-widget` updated to version `0.4.0`
+* `audit-management-widget` updated to version `0.4.0`
+* `role-management-widget` updated to version `0.3.0`
+* `user-management-widget` updated to version `0.7.0`
+* `user-profile-widget` updated to version `0.3.0`
+* `applications-portal-widget` updated to version `0.3.0`
+* `web-component` updated to version `3.36.0`
+* `web-js-sdk` updated to version `1.25.0`
+* `core-js-sdk` updated to version `2.36.0`
+## [0.9.2](https://github.com/descope/descope-js/compare/angular-sdk-0.9.1...angular-sdk-0.9.2) (2025-02-20)
+
+### Dependency Updates
+
+* `access-key-management-widget` updated to version `0.3.21`
+* `audit-management-widget` updated to version `0.3.3`
+* `role-management-widget` updated to version `0.2.25`
+* `user-management-widget` updated to version `0.6.20`
+* `user-profile-widget` updated to version `0.2.21`
+* `applications-portal-widget` updated to version `0.2.24`
+* `web-component` updated to version `3.35.1`
+* `web-js-sdk` updated to version `1.24.1`
+* `core-js-sdk` updated to version `2.35.0`
 ## [0.9.1](https://github.com/descope/descope-js/compare/angular-sdk-0.9.0...angular-sdk-0.9.1) (2025-02-12)
 
 ### Dependency Updates