+++ packages/sdks/web-component/CHANGELOG.md
@@ -2,6 +2,39 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [3.37.0](https://github.com/descope/descope-js/compare/web-component-3.36.1...web-component-3.37.0) (2025-02-26)
+
+### Dependency Updates
+
+* `sdk-mixins` updated to version `0.11.0`
+
+### Features
+
+* support components attributes ([#1030](https://github.com/descope/descope-js/issues/1030)) ([a2944a1](https://github.com/descope/descope-js/commit/a2944a13840125662b465f092cdf8663995f4769))
+
+## [3.36.1](https://github.com/descope/descope-js/compare/web-component-3.36.0...web-component-3.36.1) (2025-02-25)
+
+
+### Bug Fixes
+
+* BYOS redundant onScreenUpdate calls RELEASE ([#1031](https://github.com/descope/descope-js/issues/1031)) ([bac7a25](https://github.com/descope/descope-js/commit/bac7a25019f23ba129d1d91d7043530f73d691d3))
+
+## [3.36.0](https://github.com/descope/descope-js/compare/web-component-3.35.1...web-component-3.36.0) (2025-02-24)
+
+### Dependency Updates
+
+* `sdk-mixins` updated to version `0.10.0`
+* `web-js-sdk` updated to version `1.25.0`
+
+### Features
+
+* support cookie rename ([#1025](https://github.com/descope/descope-js/issues/1025)) RELEASE ([cc90806](https://github.com/descope/descope-js/commit/cc90806d8c97d1579d89921ee23c9bf846d11b5f))
+
+## [3.35.1](https://github.com/descope/descope-js/compare/web-component-3.35.0...web-component-3.35.1) (2025-02-20)
+
+### Dependency Updates
+
+* `web-js-sdk` updated to version `1.24.1`
 ## [3.35.0](https://github.com/descope/descope-js/compare/web-component-3.34.1...web-component-3.35.0) (2025-02-12)
 
 ### Dependency Updates